renamed modules to fit with naming conventions
