+++ ERKER2Phenopackets/src/MC4R/__init__.py
@@ -1,7 +1,7 @@
-from .MappingDicts import sex_map_erker2phenopackets, zygosity_map_erker2phenopackets,\
+from .mapping_dicts import sex_map_erker2phenopackets, zygosity_map_erker2phenopackets,\
     phenotype_status_map_erker2phenopackets
-from .ParseMC4R import parse_year_of_birth, parse_sex, parse_zygosity, parse_omim
-from .MapMC4R import map_mc4r2phenopackets, map_chunk
+from .parse_mc4r import parse_year_of_birth, parse_sex, parse_zygosity, parse_omim
+from .map_mc4r import map_mc4r2phenopackets, map_chunk
 
 __all__ = [
     'sex_map_erker2phenopackets', 'zygosity_map_erker2phenopackets',
