+++ ERKER2Phenopackets/src/MC4R/pipeline.py
@@ -9,14 +9,14 @@ import re
 
 from ERKER2Phenopackets.src.logging_ import setup_logging
 from ERKER2Phenopackets.src.utils import write_files
-from ERKER2Phenopackets.src.utils import PolarsUtils
+from ERKER2Phenopackets.src.utils import polars_utils
 from ERKER2Phenopackets.src.utils import validate
-from ERKER2Phenopackets.src.MC4R.MappingDicts import \
+from ERKER2Phenopackets.src.MC4R.mapping_dicts import \
     phenotype_label_map_erker2phenopackets
-from ERKER2Phenopackets.src.MC4R.MappingDicts import allele_label_map_erker2phenopackets
+from ERKER2Phenopackets.src.MC4R.mapping_dicts import allele_label_map_erker2phenopackets
 from ERKER2Phenopackets.src.MC4R import zygosity_map_erker2phenopackets, \
     sex_map_erker2phenopackets, phenotype_status_map_erker2phenopackets
-from ERKER2Phenopackets.src.MC4R.ParseMC4R import parse_date_of_diagnosis, \
+from ERKER2Phenopackets.src.MC4R.parse_mc4r import parse_date_of_diagnosis, \
     parse_year_of_birth, parse_phenotyping_date, parse_omim
 from ERKER2Phenopackets.src.MC4R import map_mc4r2phenopackets, map_chunk
