+++ ERKER2Phenopackets/src/__init__.py
@@ -1,5 +1,5 @@
 from . import MC4R
-from .MC4R.Pipeline import pipeline
+from .MC4R.pipeline import pipeline
 
 from . import utils
 from .utils import validate
