+++ ERKER2Phenopackets/src/logging_/__init__.py
@@ -1,4 +1,4 @@
-from .Logging_ import LOG_LEVELS, setup_logging
+from .logging_ import LOG_LEVELS, setup_logging
 
 __all__ = [
     "LOG_LEVELS", "setup_logging",
