+++ ERKER2Phenopackets/src/utils/__init__.py
@@ -1,15 +1,15 @@
-from .Phenopackets2JSON import write_phenopackets2json_files as write_files, \
+from .phenopackets2json import write_phenopackets2json_files as write_files, \
     write_phenopacket2json_file as write_file
 
-from .ParallelizationUtils import calc_chunk_size, split_dataframe
+from .parallelization_utils import calc_chunk_size, split_dataframe
 
-from .ParsingUtils import parse_date_string_to_protobuf_timestamp, \
+from .parsing_utils import parse_date_string_to_protobuf_timestamp, \
     parse_year_month_day_to_protobuf_timestamp, \
     parse_date_string_to_iso8601_utc_timestamp, \
     parse_year_month_day_to_iso8601_utc_timestamp, \
     parse_iso8601_utc_to_protobuf_timestamp
 
-from .PhenopacketValidation import validate
+from .validate_phenopackets import validate
 from .delete_files_in_folder import delete_files_in_folder
 
 __all__ = [
